+++ tests/integration/api/v2010/account/usage/record/test_yearly.py
@@ -48,6 +48,7 @@ class YearlyTestCase(IntegrationTestCase):
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "api_version": "2010-04-01",
+                        "as_of": "2019-06-24T22:32:49+00:00",
                         "category": "sms-inbound-shortcode",
                         "count": "0",
                         "count_unit": "messages",