+++ tests/integration/api/v2010/account/usage/test_record.py
@@ -47,6 +47,7 @@ class RecordTestCase(IntegrationTestCase):
                     {
                         "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "api_version": "2010-04-01",
+                        "as_of": "2019-06-24T22:32:49+00:00",
                         "category": "totalprice",
                         "count": null,
                         "count_unit": "",