+++ tests/integration/preview/acc_security/service/test_verification_check.py
@@ -37,7 +37,7 @@ class VerificationCheckTestCase(IntegrationTestCase):
                 "sid": "VEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "to": "+14159373912",
+                "to": "+15017122661",
                 "channel": "sms",
                 "status": "approved",
                 "valid": false,