+++ tests/integration/video/v1/room/room_participant/test_room_participant_subscribe_rule.py
@@ -36,7 +36,16 @@ class SubscribeRulesTestCase(IntegrationTestCase):
                 "room_sid": "RMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": null,
                 "date_updated": null,
-                "rules": []
+                "rules": [
+                    {
+                        "type": "include",
+                        "all": true,
+                        "publisher": null,
+                        "track": null,
+                        "kind": null,
+                        "priority": null
+                    }
+                ]
             }
             '''
         ))
@@ -72,7 +81,11 @@ class SubscribeRulesTestCase(IntegrationTestCase):
                 "rules": [
                     {
                         "type": "exclude",
-                        "all": true
+                        "all": true,
+                        "publisher": null,
+                        "track": null,
+                        "kind": null,
+                        "priority": null
                     }
                 ]
             }